// src/models/Element.ts
export enum Element {
	AQUARIUS = 'Aquarius',
	ARIES = 'Aries',
	RISING = 'Rising', // part of the birth chart.
	CANCER = 'Cancer',
	CAPRICORN = 'Capricorn',
	DC = 'DC',
	GEMINI = 'Gemini',
	IC = 'IC',
	JUPITER = 'Jupiter', // part of the birth chart.
	LEO = 'Leo',
	LIBRA = 'Libra',
	MARS = 'Mars', // part of the birth chart.
	MC = 'MC', // part of the birth chart.
	MERCURY = 'Mercury', // part of the birth chart.
	MOON = 'Moon', // part of the birth chart.
	NEPTUNE = 'Neptune',
	PISCES = 'Pisces',
	PLUTO = 'Pluto',
	SAGITTARIUS = 'Sagittarius',
	SATURN = 'Saturn',
	SCORPIO = 'Scorpio',
	SUN = 'Sun', // part of the birth chart.
	TAURUS = 'Taurus',
	URANUS = 'Uranus',
	VENUS = 'Venus', // part of the birth chart.
	VIRGO = 'Virgo'
}